STEP03, skip recacalculation if FULL_RECALC flag set to 0

+++ Scripts/03_FASTQC/Run_Jobs.sh
@@ -10,8 +10,19 @@ set -o nounset
 #Way to run script:
 #Usage: ./Run_Jobs.sh
 #Example: 
-#./Run_Jobs.sh
+# ./Run_Jobs.sh has one parameter - FULL_RECALC, if specify nothing by
+# default will be used FULL_RECALC=1 (full recalculation)
+# ./Run_Jobs.sh 0 -- allows to reuse previously calculated resutls
 ##################################################################################
+
+# FULL_RECALC=1 by default if nothing provided
+FULL_RECALC=${1:-1}
+
+# Skip this step if recalculation flag set to 0
+if [ ${FULL_RECALC} -eq 0 ]; then
+    echo "Recalculation is not required. FULL_RECALC set to 0."
+    exit 0
+fi
 
 # remove *.o files from previous jobs
 rm -rf ./logs
